Make python converter more flexible with maps
Now we aren't restricted to using only `tmpobjid` as the object variable and `\n` for the line endings at the cost if importing a regex library.

Also, world id is an optional parameter and some mapping tools export maps without it. This change patches it as well by using try&catch.